Make InputTriggerConfig clear/add public, move VisualEditorPanel to gui package
